+++ integration_tests/pages/staticPages/accessibilityStatement.ts
@@ -1,4 +1,4 @@
-import Page from './page'
+import Page from '../page'
 
 export default class AccessibilityStatementPage extends Page {
   constructor() {
