+++ integration_tests/pages/staticPages/privacyNotice.ts
@@ -1,4 +1,4 @@
-import Page from './page'
+import Page from '../page'
 
 export default class PrivacyNoticePage extends Page {
   constructor() {
