+++ integration_tests/pages/staticPages/signedOut.ts
@@ -1,4 +1,4 @@
-import Page, { PageElement } from './page'
+import Page, { PageElement } from '../page'
 
 export default class SignedOutPage extends Page {
   constructor() {
