+++ integration_tests/pages/staticPages/termsAndConditions.ts
@@ -1,4 +1,4 @@
-import Page from './page'
+import Page from '../page'
 
 export default class TermsAndConditionsPage extends Page {
   constructor() {
