+++ server/routes/authenticatedRoutes.ts
