+++ server/routes/cookies/cookiesController.test.ts
@@ -2,9 +2,9 @@ import type { Express } from 'express'
 import request from 'supertest'
 import * as cheerio from 'cheerio'
 import { FieldValidationError } from 'express-validator'
-import { appWithAllRoutes, FlashData, flashProvider } from './testutils/appSetup'
-import paths from '../constants/paths'
-import config from '../config'
+import { appWithAllRoutes, FlashData, flashProvider } from '../testutils/appSetup'
+import paths from '../../constants/paths'
+import config from '../../config'
 
 let app: Express
