+++ server/routes/unauthenticatedRoutes.ts
@@ -1,8 +1,9 @@
 import { Router } from 'express'
+import type { Services } from '../services'
 import paths from '../constants/paths'
-import CookiesController from './cookiesController'
-import { Services } from '../services'
 import config from '../config'
+import CookiesController from './cookies/cookiesController'
+import staticPagesRoutes from './staticPages'
 
 export default function routes({ prisonService }: Services): Router {
   const router = Router()
@@ -22,30 +23,13 @@ export default function routes({ prisonService }: Services): Router {
     return res.render('pages/serviceStart', { supportedPrisons, hideGOVUKServiceNav: true })
   })
 
-  // Accessibility statement
-  router.get(paths.ACCESSIBILITY, (req, res) => {
-    res.render('pages/accessibilityStatement', { showOLServiceNav: !!req.session.booker })
-  })
-
   // Cookies
   const cookies = new CookiesController()
   router.get(paths.COOKIES, cookies.view())
   router.post(paths.COOKIES, cookies.validate(), cookies.submit())
 
-  // Privacy notice
-  router.get(paths.PRIVACY, (req, res) => {
-    res.render('pages/privacyNotice', { showOLServiceNav: !!req.session.booker })
-  })
-
-  // Terms and conditions
-  router.get(paths.TERMS, (req, res) => {
-    res.render('pages/termsAndConditions', { showOLServiceNav: !!req.session.booker })
-  })
-
-  // Signed out
-  router.get(paths.SIGNED_OUT, (req, res) => {
-    res.render('pages/signedOut')
-  })
+  // Static pages
+  router.use(staticPagesRoutes())
 
   return router
 }
